test: extract multi-color vector PDF fixture for non-browser gate
Keep the colors>1 E2E assertion, but move the fixture helper so the
deliberate-break check runs without Playwright installed locally.
